Fix standalone wheel artifact build
In order to run on Dataflow, a standalone wheel needs to be uploaded.
Wheel uploading did not work out of the box due to assumptions made
by the grpcio_tools setuptools command. This change ensures that all
generated files are part of the wheel distribution, without breaking
the local install path.

Change-Id: I96803363ea5e612ed43c38496ca61bee84371669